pdf-expenses-parser/requirements.txt
azure-functions
pdfplumber
pandas
